+++ extensions/norwegian-public-transport/src/Search/Actions.tsx
@@ -1,4 +1,4 @@
-import { Action, ActionPanel, Icon, Image } from "@raycast/api";
+import { Action, ActionPanel, Icon, Image, Keyboard } from "@raycast/api";
 import { getFavicon } from "@raycast/utils";
 import { Feature } from "../types";
 
@@ -16,13 +16,11 @@ export function Actions({ venue, isFavorite, onAction, onSave }: ActionsProps) {
       <Action
         title={
           isFavorite
-            ? `Remove ${venue.properties.name} From Favorites`
+            ? `Remove ${venue.properties.name} from Favorites`
             : `Add ${venue.properties.name} to Favorites`
         }
         icon={isFavorite ? Icon.StarDisabled : Icon.Star}
-        shortcut={
-          isFavorite ? { modifiers: ["cmd", "shift"], key: "s" } : { modifiers: ["cmd"], key: "s" }
-        }
+        shortcut={Keyboard.Shortcut.Common.Pin}
         onAction={onSave}
       />
       {venue.properties.id && (